Remove `trans` attribute from .tmx files
This was breaking transparency for some reason. Details in #46
but copied here for historical reasons:

Bitcraft:
i took a look at your transparency issue and found it. it was an odd
one... if you look at the tmx level files, you'll notice that the
tileset images have a attribute called "trans" set.

  <image source="../../newtiledsheet.png" trans="ff00ff" width="144"
  height="224"/> this is supposed to mean "colorkey". what is happening
  here is that the maps are loaded with pytmx as 32-bit images with an
  alpha channel. pyscroll will reduce all the tiles when loaded so it
  draws faster, if possible. because tiled has a colorkey set (the trans
  attribute), pyscroll tries to treat the tiles as having a color key,
  so the alpha channel is lost and leaves some tiles with an ugly
  border.

the alpha kwarg for the bufferered renderer isn't obvious. all tiles
will have an alpha channel if they require it, but the layers do not.
this flag means "give the layers an alpha channel". enabling reduces
performance by about 30% because there are 8 more bits to write per
pixel for each layer, so it is disabled by default. it is only needed
for parallax scrolling and other effects.

TL;DR - your tileset is misconfigured to use a color key when it really
has a alpha channel

edit the tmx files and remove the color key attribute ("trans") remove
the "alpha=True" kwarg from the pyscroll buffered renderer