Make pressure plates work in real-time
Because of how mr-figs works (it's turn based), it's difficult to do
collision detection per turn. As a workaround we check for pressure
plate collisions *per frame* just like a normal game.

All other collisions are done per turn (since it makes sense and is more
efficient). I think there's probably still a way of changing this to be
based on a turn instead of frame if we can wrangle the movement_vector
and the player's direction we can probably get it working. Look at the
moveable_tile logic as an example for inspiration.