Fix buggy "left" behaviour
We had the speed of the player's movement set to 6 (Not an even divisor
of 16 - the width of the tiles) which would then be interpolated to (I
think). It only affected the left because I only had interpolation set
on that direction.

Instead, just divide the tile_width and be done with it.