Refactor `solid` use
- We have removed as much of a mention of `solid` as possible. We should
just be able to use the new `SolidTile` and not rely on `solid`
attributes.

- We have renamed `Tile` to `SolidTile` to make it more descriptive of
what it's actually doing

- We have refactored bomb particle creation. Whereas before the
responsiblity for where a particle was created lied within the
`Bomb` class, it now lies with the respective tile. This simplifies the
logic within `bomb.py` and also means we can have special behaviour per
tile if we wish. To accommodate this we've added both a "pre" and "post"
method for when a particle attempts to get created on a tile