Create test.sh
I think acceptance tests are the way to go for this. We're deeply
entwined with pygame and it's difficult not to be.

In the future I'd like to see test.sh as the wrapper to run all other
test files. Each test should be a separate file and also a separate
level in Tiled.