Refactor to use pygame.sprite.Sprite:update()
Okay so a fairly small change that could have big impact.

We've changed it so that we now call `pygame.sprite.Sprite:update()`
within the main game loop. This has the benefit of us just being able to
add an `update()` method to any sprite we want to add some logic to. The
downside is that I was already using `update()` as a method name on a
few classes. I've updated these to different name or removed them
entirely so I *think* we're bug free.

This refactor as has brought up some issues with the triggerable tile
now. After this refactor, the triggerable now has to be above the
corresponding pressure_plate in Tiled. No idea why, probably something
to do with state and changing its image. That's the only bug I've found
so far.