Add basic xbox 360 controller support
This adds support in some areas of the game for xbox 360 controllers.
In the future it'd be good to make it even more generic so we don't need
both a `process_input` and a `process_joystick_input`. Ideally they
should work the same. I have created a `KeyboardController` class to
help achieve this goal. Once they are all behind a common interface we
can tidy it up even more.