Refactor all events to be unified
Big change. This will listen to pygame's event queue and then wrap any
events into a unified way. This means that a press on a keyboard, or a
controller, or a mouse behave in exactly the same way.

This is separate to the `Controller` classes that we have which are
simply mappings of the input for each device.

In doing this we have massively simplified the input handling but also
lost some flexibility too. We no longer have hotkeys for each menu item.
Now, we use the up/down and action keys (on whatever device that may be)
to select menu items.

We've also updated the options menu as this was out of date and not
using the newest `MenuItems` code.