Fix laser collision on moving_laser
The biggest issue here was not clearing the laser sprites immediately on
every `update`. I think this lead to us checking stale lasers in some
cases so the rendering and collision was off.

We've added a `solids` group to the level too to speed things up.
Hopefully we can use this elsewhere too.