Implement player hitboxes
Now that Mr Figs has an offset and is spanning more than just one tile,
we need to implement a hitbox so that the collision at least feels fair.

Basically we attach an entity to him (`collideable`) and we should now
check all rect collisions against that instead of the player rect.